1654 Move finish confirmation message hook to finish run
